Global Tate models

Constructors

Attributes

Global Tate models without specified base space

This method constructs a global Tate model over a base space that is not fully specified. Rather, it assums that a base space exists such that the Tate sections $a_i$ are well-defined so that the Tate model in question is well-defined.

For many practical applications, one wishes to assume a further factorization of the Tate sections $a_i$. This has the advantage that one can engineer singularity loci or even the singularity type over a specific locus. This is the backbone of many F-theory constructions.

To this end, this method accepts a polynomial ring whose variables are the sections used in the desired factorization of the Tate sections $a_i$. For example, if we desired a factorization:

$a_1 = a_{10} w^0$,

$a_2 = a_{21} w^1$,

$a_3 = a_{32} w^2$,

$a_4 = a_{43} w^3$,

$a_6 = a_{65} w^5$,

then the polynomial ring in question is the ring with indeterminates $a_{10}$, $a_{21}$, $a_{32}$, $a_{43}$, $a_{65}$ and $w$.

In theory, one can consider these indeterminates as local coordinate of the base space. For the computer implementation, this ring will therefore serve as the coordinate ring of an auxiliary toric base space, namely an affine space with those coordinates.

For such geometries, we support the following functionality.

Constructors

FTheoryTools.GlobalTateModel — Method

GlobalTateModel(ais::Vector{fmpq_mpoly}, auxiliary_base_ring::MPolyRing)

This method constructs a global Tate model over a base space that is not fully specified. The following example exemplifies this approach.

Examples

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> tate_polynomial(t)
a10*x*y*z + a21*w*x^2*z^2 + a32*w^2*y*z^3 + a43*w^3*x*z^4 + a65*w^5*z^6 + x^3 - y^2

julia> auxiliary_base_space(t)
A normal, affine, 6-dimensional toric variety

julia> auxiliary_ambient_space(t)
A normal toric variety

julia> dim(auxiliary_ambient_space(t))
8

Attributes

FTheoryTools.tate_section_a1 — Method

tate_section_a1(t::TateModelOverGeneralBaseSpace)

Return the Tate section $a_1$.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> tate_section_a1(t)
a10

FTheoryTools.tate_section_a2 — Method

tate_section_a2(t::TateModelOverGeneralBaseSpace)

Return the Tate section $a_2$.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> tate_section_a2(t)
a21*w

FTheoryTools.tate_section_a3 — Method

tate_section_a3(t::TateModelOverGeneralBaseSpace)

Return the Tate section $a_3$.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> tate_section_a3(t)
a32*w^2

FTheoryTools.tate_section_a4 — Method

a4(t::TateModelOverGeneralBaseSpace)

Return the Tate section $a_4$.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> tate_section_a4(t)
a43*w^3

FTheoryTools.tate_section_a6 — Method

tate_section_a6(t::TateModelOverGeneralBaseSpace)

Return the Tate section $a_6$.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> tate_section_a6(t)
a65*w^5

FTheoryTools.tate_polynomial — Method

tate_polynomial(t::TateModelOverGeneralBaseSpace)

Return the Tate polynomial of the global Tate model.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> tate_polynomial(t)
a10*x*y*z + a21*w*x^2*z^2 + a32*w^2*y*z^3 + a43*w^3*x*z^4 + a65*w^5*z^6 + x^3 - y^2

FTheoryTools.auxiliary_base_space — Method

auxiliary_base_space(t::TateModelOverGeneralBaseSpace)

Return the toric base space of the global Tate model.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> auxiliary_base_space(t)
A normal, affine, 6-dimensional toric variety

FTheoryTools.auxiliary_ambient_space — Method

auxiliary_ambient_space(t::TateModelOverGeneralBaseSpace)

Return the toric ambient space of the global Tate model.

julia> using Oscar

julia> auxiliary_base_ring, (a10,a21,a32,a43,a65,w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = GlobalTateModel(ais, auxiliary_base_ring)
A global Tate model over a not fully specified base

julia> auxiliary_ambient_space(t)
A normal toric variety

julia> dim(auxiliary_ambient_space(t))
8