Global Weierstrass models

Constructors

Attributes

Global Weierstrass models

This method constructs a global Weierstrass model over a base space that is not fully specified. Rather, it assums that a base space exists such that the Weierstrass sections $f$ and $g$ are well-defined, so that the Weierstrass model in question is well-defined.

For many practical applications, one wishes to assume a further specialize the Weierstrass sections $f$ and $g$. This has the advantage that one can engineer singularity loci or even the singularity type over a specific locus. To some extend, this is the backbone of many F-theory constructions.

Consequently, the construction of such models accepts a polynomial ring whose variables are the sections used in the desired factorization of the Weierstrass sections $f$ and $g$.

In theory, one can consider the indeterminates of this auxiliary ring as a (possilby redundant) set of local coordinate of the base space. For the computer implementation, this ring will therefore serve as the coordinate ring of an auxiliary toric base space, namely an affine space with those coordinates.

For such geometries, we support the following functionality.

Constructors

We support the following constructors:

FTheoryTools.GlobalWeierstrassModel — Method

GlobalWeierstrassModel(f::fmpq_mpoly, g::fmpq_mpoly, auxiliary_base_ring::MPolyRing)

This method constructs a global Weierstrass model over a base space that is not fully specified. The following example illustrates this approach.

Examples

julia> using Oscar

julia> auxiliary_base_ring, (f, g) = QQ["f", "g"];

julia> w = GlobalWeierstrassModel(f, g, auxiliary_base_ring)
A global Weierstrass model over a not fully specified base

julia> weierstrass_polynomial(w)
f*x*z^4 + g*z^6 + x^3 - y^2

julia> auxiliary_base_space(w)
A normal, affine, 2-dimensional toric variety

julia> auxiliary_ambient_space(w)
A normal toric variety

julia> dim(auxiliary_ambient_space(w))
4

Attributes

FTheoryTools.weierstrass_section_f — Method

weierstrass_section_f(w::WeierstrassModelOverGeneralBaseSpace)

Return the Weierstrass section $f$.

julia> using Oscar

julia> auxiliary_base_ring, (f, g) = QQ["f", "g"];

julia> w = GlobalWeierstrassModel(f, g, auxiliary_base_ring)
A global Weierstrass model over a not fully specified base

julia> weierstrass_section_f(w)
f

FTheoryTools.weierstrass_section_g — Method

weierstrass_section_g(w::WeierstrassModelOverGeneralBaseSpace)

Return the Weierstrass section $g$.

julia> using Oscar

julia> auxiliary_base_ring, (f, g) = QQ["f", "g"];

julia> w = GlobalWeierstrassModel(f, g, auxiliary_base_ring)
A global Weierstrass model over a not fully specified base

julia> weierstrass_section_g(w)
g

FTheoryTools.weierstrass_polynomial — Method

weierstrass_polynomial(w::WeierstrassModelOverGeneralBaseSpace)

Return the Weierstrass polynomial.

julia> using Oscar

julia> auxiliary_base_ring, (f, g) = QQ["f", "g"];

julia> w = GlobalWeierstrassModel(f, g, auxiliary_base_ring)
A global Weierstrass model over a not fully specified base

julia> weierstrass_polynomial(w)
f*x*z^4 + g*z^6 + x^3 - y^2

AbstractAlgebra.discriminant — Method

discriminant(w::WeierstrassModelOverGeneralBaseSpace)

Return the discriminant $\Delta = 4 f^3 + 27 g^2$.

julia> using Oscar

julia> auxiliary_base_ring, (f, g) = QQ["f", "g"];

julia> w = GlobalWeierstrassModel(f, g, auxiliary_base_ring)
A global Weierstrass model over a not fully specified base

julia> discriminant(w);

FTheoryTools.auxiliary_base_space — Method

auxiliary_base_space(w::WeierstrassModelOverGeneralBaseSpace)

Return the toric base space of the global Tate model.

julia> using Oscar

julia> auxiliary_base_ring, (f, g) = QQ["f", "g"];

julia> w = GlobalWeierstrassModel(f, g, auxiliary_base_ring)
A global Weierstrass model over a not fully specified base

julia> auxiliary_base_space(w)
A normal, affine, 2-dimensional toric variety

FTheoryTools.auxiliary_ambient_space — Method

auxiliary_ambient_space(w::WeierstrassModelOverGeneralBaseSpace)

Return the toric ambient space of the global Tate model.

julia> using Oscar

julia> auxiliary_base_ring, (f, g) = QQ["f", "g"];

julia> w = GlobalWeierstrassModel(f, g, auxiliary_base_ring)
A global Weierstrass model over a not fully specified base

julia> auxiliary_ambient_space(w)
A normal toric variety