Global Tate models

Constructors

Attributes

Global Tate models

A global Tate model describes a particular form of an elliptic fibration. We focus on elliptic fibrations over base 3-folds $B3$. Consider the weighted projective space $\mathbb{P}^{2,3,1}$ with coordinates $x, y, z$. In addition, consider

$a_1 \in H^0( B_3, \overline{K}_{B_3} )$,

$a_2 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 2} )$,

$a_3 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 3} )$,

$a_4 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 4} )$,

$a_6 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 6} )$.

Then form a $\mathbb{P}^{2,3,1}$-bundle over $B3$ such that

$x$ transforms as a section of $2 \overline{K}_{B_3}$,

$y$ transforms as a section of $3 \overline{K}_{B_3}$,

$z$ transforms as a section of $0 \overline{K}_{B_3} = \mathcal{O}_{B_3}$.

In this 5-fold ambient space, a global Tate model is the hypersurface defined by the vanishing of the Tate polynomial $P_T = x^3 - y^2 - x y z a_1 + x^2 z^2 a_2 - y z^3 a_3 + x z^4 a_4 + z^6 a_6$. For such geometries, we support functionality.

Constructors

FTheoryTools.GenericGlobalTateModel — Method

GenericGlobalTateModel(base::Oscar.AbstractNormalToricVariety)

This method constructs a global Tate model over a given toric base 3-fold. The Tate sections $a_i$ are taken with (pseudo) random coefficients.

One way to achieve this is to first focus on the Cox ring of the toric ambient space. This ring must be graded such that the Tate polynomial is homogeneous and cuts out a Calabi-Yau hypersurface. Given this grading, one can perform a triangulation. This triangulation will typically take a long time to complete and yield a large number of candidate ambient spaces. Typically, one wishes to focus on those spaces which contain the base toric space in a manifest way. But even this criterion usually allows for many ambient spaces.

This method here operates in the opposite way. It begins by extracting the rays and maximal cones of the base space. Subsequently, those rays and cones are extended to form one of the many toric 5-fold ambient spaces. The following example demonstrates that this ambient space of a (singular) global Tate model need not be smooth.

Examples

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> is_smooth(toric_ambient_space(t))
false

FTheoryTools.GenericGlobalTateModelOverProjectiveSpace — Method

GenericGlobalTateModelOverProjectiveSpace()

This method constructs a global Tate model over the 3-dimensional projective space.

Examples

julia> using Oscar

julia> GenericGlobalTateModelOverProjectiveSpace()
A global Tate model over a concrete base

FTheoryTools.SpecificGlobalTateModel — Method

SpecificGlobalTateModel(ais::Vector{MPolyElem_dec{fmpq, fmpq_mpoly}}, base::Oscar.AbstractNormalToricVariety)

This method operates analogously to GenericGlobalTateModel(base::Oscar.AbstractNormalToricVariety). The only difference is that the Tate sections $a_i$ can be specified with non-generic values.

Examples

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> a1 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base))]);

julia> a2 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^2)]);

julia> a3 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^3)]);

julia> a4 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^4)]);

julia> a6 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^6)]);

julia> t = SpecificGlobalTateModel([a1, a2, a3, a4, a6], base)
A global Tate model over a concrete base

julia> is_smooth(toric_ambient_space(t))
false

Attributes

FTheoryTools.tate_section_a1 — Method

tate_section_a1(t::GlobalTateModel)

Return the Tate section $a_1$.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> tate_section_a1(t);

FTheoryTools.tate_section_a2 — Method

tate_section_a2(t::GlobalTateModel)

Return the Tate section $a_2$.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> tate_section_a2(t);

FTheoryTools.tate_section_a3 — Method

tate_section_a3(t::GlobalTateModel)

Return the Tate section $a_3$.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> tate_section_a3(t);

FTheoryTools.tate_section_a4 — Method

tate_section_a4(t::GlobalTateModel)

Return the Tate section $a_4$.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> tate_section_a4(t);

FTheoryTools.tate_section_a6 — Method

tate_section_a6(t::GlobalTateModel)

Return the Tate section $a_6$.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> tate_section_a6(t);

FTheoryTools.tate_polynomial — Method

tate_polynomial(t::GlobalTateModel)

Return the Tate polynomial of the global Tate model.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> tate_polynomial(t);

FTheoryTools.toric_base_space — Method

toric_base_space(t::GlobalTateModel)

Return the toric base space of the global Tate model.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> toric_base_space(t)
A normal, 3-dimensional toric variety without torusfactor

FTheoryTools.toric_ambient_space — Method

toric_ambient_space(t::GlobalTateModel)

Return the toric ambient space of the global Tate model.

julia> using Oscar

julia> test_space = hirzebruch_surface(2) * projective_space(NormalToricVariety,1)
A normal toric variety

julia> test_space1 = blowup_on_ith_minimal_torus_orbit(test_space,1,"e1")
A normal toric variety

julia> test_space2 = blowup_on_ith_minimal_torus_orbit(test_space1,1,"e2")
A normal toric variety

julia> base = blowup_on_ith_minimal_torus_orbit(test_space2,1,"e3")
A normal toric variety

julia> t = GenericGlobalTateModel(base)
A global Tate model over a concrete base

julia> toric_ambient_space(t)
A normal toric variety

julia> is_smooth(toric_ambient_space(t))
false