Global Weierstrass models

Constructors

Attributes

Global Weierstrass models

Constructors

We support the following constructors:

Missing docstring.

Missing docstring for GlobalWeierstrassModel(polys::Vector{MPolyElem{fmpq}}). Check Documenter's build log for details.

FTheoryTools.GenericGlobalWeierstrassModelOverToricSpace — Method

GenericGlobalWeierstrassModelOverToricSpace(v::Oscar.AbstractNormalToricVariety)

Given a toric variety $v$, this method determines generic sections $f \in H^0( v, \overline{K}_v^4)$ and $g \in H^0( v, \overline{K}_v^6)$. Subsequently, it constructs the global Weierstrass model defined by those two polynomials.

FTheoryTools.GenericGlobalWeierstrassModelOverProjectiveSpace — Method

GenericGlobalWeierstrassModelOverProjectiveSpace(n::Int)

This method constructs the $n$-dimensional projective space, determines generic sections $f$ and $g$ of the 4-th and 6-th power of the anticanonical bundle and then constructs the global Weierstrass model defined by those two polynomials.

julia> w = GenericGlobalWeierstrassModelOverProjectiveSpace(3)
A global Weierstrass model

Attributes

FTheoryTools.poly_f — Method

poly_f(w::GlobalWeierstrassModel)

Return the polynomial $f$ used for the construction of the global Weierstrass model.

julia> w = GenericGlobalWeierstrassModelOverProjectiveSpace(3)
A global Weierstrass model

julia> poly_f(w);

FTheoryTools.poly_g — Method

poly_g(w::GlobalWeierstrassModel)

Return the polynomial $g$ used for the construction of the global Weierstrass model.

julia> w = GenericGlobalWeierstrassModelOverProjectiveSpace(3)
A global Weierstrass model

julia> poly_f(w);

FTheoryTools.toric_base_space — Method

toric_base_space(w::GlobalWeierstrassModel)

If the global Weierstrass model in question was constructed over a toric base space, this method returns this toric space. Otherwise an error is raised.

julia> w = GenericGlobalWeierstrassModelOverProjectiveSpace(3)
A global Weierstrass model

julia> toric_base_space(w)
A normal, non-affine, smooth, projective, gorenstein, fano, 3-dimensional toric variety without torusfactor